benchmarks/requirements.txt
numpy
usearch
faiss-cpu
matplotlib
